refactor(temperature): 気温計算をqueryへ分離し入眠まわりを整える
states がUI層から systems を直接呼ぶ違和感を解消する。気温計算の dungeon 依存は
「基本気温 int を1つ引く」だけだったので、生成時に StageField.BaseTemp へ写し、
気温計算 AmbientTemperatureAt などを query へ移す。query→dungeon の import cycle を
避けつつ純粋計算を基盤層へ置く。EvaluateSleepConditions は sleep behavior と同じ
activity へ集約し、states→activity にする。TemperatureSystem は systems に残す。

あわせて入眠のゲートを SleepBehavior.Validate に一本化し、疲労・気温・安全を同じ
EvaluateSleepConditions で判定してプロンプト表示とずらさない。

継続アクティビティの中断まわりのバグも直す。DoTurn の自己中断で Canceled フックが
走らず起床処理と理由が出ていなかったのを stepActivity で処理する。CancelActivity は
理由設定を Canceled より前に移す。Canceled は Sleeping 除去で comp が無効化される前に
理由を退避する。低体温での起床は状態の存在でなく IsActive で判定し、火で回復中の残留では
起こさない。

テスト: 疲労テスト名を挙動が伝わる形へ改名、睡眠中の空腹半減テストを追加、
温度テストを query/systems へ分割する。
