+++ internal/components/stage_field.go
@@ -11,6 +11,10 @@ type StageField struct {
 	// Level は現ステージのフィールド寸法。ステージごとに保持するため、往復してもステージ固有の
 	// 寸法が resume で自然に戻る。
 	Level Level
+	// BaseTemp はステージの基本気温。摂氏。ステージ生成時に dungeon 定義から確定して写す。
+	// 気温計算を query 層に置くため、dungeon 登録表への依存をここで断つ。登録表は別層なので
+	// query からは読めず、Level と同じくステージ生成時に定まる設定として保持する。
+	BaseTemp int
 	// ExploredTiles は探索済みタイルのマップ。ステージごとに保持する。
 	// GridElement(struct)キーのためserde不可、入場時リセット方針なのでロード後は空で再構築する
 	ExploredTiles map[GridElement]bool `json:"-"`
