+++ internal/systems/temperature_test.go
@@ -5,7 +5,9 @@ import (
 
 	gc "github.com/kijimaD/ruins/internal/components"
 	"github.com/kijimaD/ruins/internal/consts"
+	"github.com/kijimaD/ruins/internal/dungeon"
 	"github.com/kijimaD/ruins/internal/testutil"
+	w "github.com/kijimaD/ruins/internal/world"
 
 	"github.com/kijimaD/ruins/internal/world/lifecycle"
 	"github.com/kijimaD/ruins/internal/world/query"
@@ -16,201 +18,11 @@ import (
 // coldDungeonName は基本気温0度のテスト用ダンジョン定義名。DungeonForest の英語 id。
 const coldDungeonName = "Dead forest"
 
-func TestTileEnvironmentAt(t *testing.T) {
-	t.Parallel()
-
-	t.Run("囲われは緩和度として返り加算℃には入らない", func(t *testing.T) {
-		t.Parallel()
-		world := testutil.InitTestWorld(t)
-
-		tile := world.ECS.NewEntity()
-		world.Components.GridElement.Add(tile, &gc.GridElement{Coord: consts.Coord[consts.Tile]{X: 5, Y: 5}})
-		world.Components.TileEnvironment.Add(tile, &gc.TileEnvironment{
-			Shelter: gc.ShelterFull,
-		})
-
-		shelter, modifier := tileEnvironmentAt(world, 5, 5)
-		assert.Equal(t, gc.ShelterFull, shelter)
-		assert.Equal(t, 0, modifier)
-	})
-
-	t.Run("水と植生は加算℃として返る", func(t *testing.T) {
-		t.Parallel()
-		world := testutil.InitTestWorld(t)
-
-		tile := world.ECS.NewEntity()
-		world.Components.GridElement.Add(tile, &gc.GridElement{Coord: consts.Coord[consts.Tile]{X: 5, Y: 5}})
-		world.Components.TileEnvironment.Add(tile, &gc.TileEnvironment{
-			Water:   gc.WaterNearby,
-			Foliage: gc.FoliageForest,
-		})
-
-		shelter, modifier := tileEnvironmentAt(world, 5, 5)
-		assert.Equal(t, gc.ShelterNone, shelter)
-		assert.Equal(t, -8, modifier)
-	})
-
-	t.Run("タイルが存在しない場合は屋外かつ0を返す", func(t *testing.T) {
-		t.Parallel()
-		world := testutil.InitTestWorld(t)
-
-		shelter, modifier := tileEnvironmentAt(world, 5, 5)
-		assert.Equal(t, gc.ShelterNone, shelter)
-		assert.Equal(t, 0, modifier)
-	})
-}
-
-func TestAmbientTemperatureAt_オーバーワールドは季節の世界温度そのもの(t *testing.T) {
-	t.Parallel()
-	world := testutil.InitTestWorld(t)
-	// 25日目の昼。冬の底 -30 に昼補正 +10
-	query.GetGameTime(world).TotalTurns = 24*1500 + 500
-	// 帯データの有無が屋外判定を兼ねる。オーバーワールドにして帯を付ける
-	query.GetDungeon(world).CurrentStage = gc.NewOverworldStage()
-	query.EnsureSeamlessBand(world)
-
-	temp, err := AmbientTemperatureAt(world, 0, 0)
-	require.NoError(t, err)
-	assert.Equal(t, -20, temp, "屋外は季節世界温度 -30 に昼補正 +10 を足す")
-}
-
-func TestAmbientTemperatureAt_屋外はステージ定義が無くても世界温度を返す(t *testing.T) {
-	t.Parallel()
-	world := testutil.InitTestWorld(t)
-	// 25日目の昼。世界温度 = -30 + 10 = -20
-	query.GetGameTime(world).TotalTurns = 24*1500 + 500
-	// 定義が登録されていない屋外ステージ。帯を付けて屋外と判定させる
-	query.GetDungeon(world).CurrentStage = gc.StageKey{Name: "未登録の屋外"}
-	query.EnsureSeamlessBand(world)
-
-	temp, err := AmbientTemperatureAt(world, 0, 0)
-	require.NoError(t, err)
-	assert.Equal(t, -20, temp, "屋外はステージ定義に依らず世界温度そのものを返す")
-}
-
-func TestAmbientTemperatureAt_ダンジョンの囲われたタイルは世界温度を緩和して受ける(t *testing.T) {
-	t.Parallel()
-	world := testutil.InitTestWorld(t)
-	// 25日目の昼。世界温度 = -30 + 10 = -20。屋内は基本気温0に世界温度の半分を足す
-	query.GetGameTime(world).TotalTurns = 24*1500 + 500
-	query.GetDungeon(world).CurrentStage = gc.NewDungeonStage(coldDungeonName, 1)
-
-	tile := world.ECS.NewEntity()
-	world.Components.GridElement.Add(tile, &gc.GridElement{Coord: consts.Coord[consts.Tile]{X: 0, Y: 0}})
-	world.Components.TileEnvironment.Add(tile, &gc.TileEnvironment{Shelter: gc.ShelterFull})
-
-	temp, err := AmbientTemperatureAt(world, 0, 0)
-	require.NoError(t, err)
-	assert.Equal(t, -5, temp, "屋内はアンカー10へ半分寄せた 10+(-20-10)/2 = -5 になる")
-}
-
-func TestAmbientTemperatureAt_ダンジョンの囲われていないタイルは世界温度をそのまま受ける(t *testing.T) {
-	t.Parallel()
-	world := testutil.InitTestWorld(t)
-	// 25日目の昼。世界温度 = -30 + 10 = -20。洞窟の開口部のような囲われの無いタイルは屋外扱い
-	query.GetGameTime(world).TotalTurns = 24*1500 + 500
-	query.GetDungeon(world).CurrentStage = gc.NewDungeonStage(coldDungeonName, 1)
-
-	temp, err := AmbientTemperatureAt(world, 0, 0)
-	require.NoError(t, err)
-	assert.Equal(t, -20, temp, "囲われが無ければ基本気温 0 に世界温度 -20 をそのまま足す")
-}
-
-func TestAmbientTemperatureAt_オーバーワールドの屋内タイルは世界温度を緩和して受ける(t *testing.T) {
-	t.Parallel()
-	world := testutil.InitTestWorld(t)
-	// 25日目の昼。世界温度 = -30 + 10 = -20。建物内部の床は ShelterFull が設定されている
-	query.GetGameTime(world).TotalTurns = 24*1500 + 500
-	query.GetDungeon(world).CurrentStage = gc.NewOverworldStage()
-	query.EnsureSeamlessBand(world)
-
-	tile := world.ECS.NewEntity()
-	world.Components.GridElement.Add(tile, &gc.GridElement{Coord: consts.Coord[consts.Tile]{X: 0, Y: 0}})
-	world.Components.TileEnvironment.Add(tile, &gc.TileEnvironment{Shelter: gc.ShelterFull})
-
-	temp, err := AmbientTemperatureAt(world, 0, 0)
-	require.NoError(t, err)
-	assert.Equal(t, -5, temp, "オーバーワールドの屋内はアンカー10へ半分寄せた -5 になる")
-}
-
-func TestAmbientTemperatureAt_温暖時も屋内は屋外より寒くならない(t *testing.T) {
-	t.Parallel()
-	world := testutil.InitTestWorld(t)
-	// 開始直後の春の夜明け。世界温度 = 5 + 0 = 5。0℃へ割るだけの式だと屋内 2℃ で屋外より寒い逆転が起きる
-	query.GetGameTime(world).TotalTurns = 0
-	query.GetDungeon(world).CurrentStage = gc.NewOverworldStage()
-	query.EnsureSeamlessBand(world)
-
-	tile := world.ECS.NewEntity()
-	world.Components.GridElement.Add(tile, &gc.GridElement{Coord: consts.Coord[consts.Tile]{X: 0, Y: 0}})
-	world.Components.TileEnvironment.Add(tile, &gc.TileEnvironment{Shelter: gc.ShelterFull})
-
-	temp, err := AmbientTemperatureAt(world, 0, 0)
-	require.NoError(t, err)
-	assert.Equal(t, 8, temp, "屋内はアンカー10へ半分寄せた 10+(5-10)/2 = 8 で屋外の 5 より暖かい")
-}
-
-func TestAmbientTemperatureAt_熱源は環境気温を押し上げる(t *testing.T) {
-	t.Parallel()
-	world := testutil.InitTestWorld(t)
-	// 25日目の昼。世界温度 = -30 + 10 = -20。焚き火と同じ warmth 0.75 の熱源の隣に立つ
-	query.GetGameTime(world).TotalTurns = 24*1500 + 500
-	query.GetDungeon(world).CurrentStage = gc.NewOverworldStage()
-	query.EnsureSeamlessBand(world)
-
-	fire := world.ECS.NewEntity()
-	world.Components.GridElement.Add(fire, &gc.GridElement{Coord: consts.Coord[consts.Tile]{X: 0, Y: 0}})
-	world.Components.HeatSource.Add(fire, &gc.HeatSource{Radius: 2, Warmth: 0.75})
-
-	temp, err := AmbientTemperatureAt(world, 1, 0)
-	require.NoError(t, err)
-	assert.Equal(t, -5, temp, "隣接の押し上げは 0.75*2/3*30=15℃ で -20 が -5 になる")
-}
-
-func TestAmbientTemperatureAt_半屋外タイルは世界温度を中間の強さで受ける(t *testing.T) {
-	t.Parallel()
-	world := testutil.InitTestWorld(t)
-	// 25日目の昼。世界温度 = -30 + 10 = -20。半屋外は 3/4 の強さで受ける
-	query.GetGameTime(world).TotalTurns = 24*1500 + 500
-	query.GetDungeon(world).CurrentStage = gc.NewOverworldStage()
-	query.EnsureSeamlessBand(world)
-
-	tile := world.ECS.NewEntity()
-	world.Components.GridElement.Add(tile, &gc.GridElement{Coord: consts.Coord[consts.Tile]{X: 0, Y: 0}})
-	world.Components.TileEnvironment.Add(tile, &gc.TileEnvironment{Shelter: gc.ShelterPartial})
-
-	temp, err := AmbientTemperatureAt(world, 0, 0)
-	require.NoError(t, err)
-	assert.Equal(t, -12, temp, "半屋外はアンカー10へ 3/4 寄せた 10+(-30)*3/4 = -12 になる")
-}
-
-func TestAmbientTemperatureAt_冬の屋内は屋外より暖かい(t *testing.T) {
-	t.Parallel()
-	// 冬の底の同じ時刻で、屋内と屋外の周囲気温を比べる。屋内が屋外より暖かく寒さの逆転がない
-	const winterNoon consts.Turn = 24*1500 + 500
-
-	outdoor := func() int {
-		world := testutil.InitTestWorld(t)
-		query.GetGameTime(world).TotalTurns = winterNoon
-		query.GetDungeon(world).CurrentStage = gc.NewOverworldStage()
-		query.EnsureSeamlessBand(world)
-		temp, err := AmbientTemperatureAt(world, 0, 0)
-		require.NoError(t, err)
-		return temp
-	}()
-	indoor := func() int {
-		world := testutil.InitTestWorld(t)
-		query.GetGameTime(world).TotalTurns = winterNoon
-		query.GetDungeon(world).CurrentStage = gc.NewDungeonStage(coldDungeonName, 1)
-		tile := world.ECS.NewEntity()
-		world.Components.GridElement.Add(tile, &gc.GridElement{Coord: consts.Coord[consts.Tile]{X: 0, Y: 0}})
-		world.Components.TileEnvironment.Add(tile, &gc.TileEnvironment{Shelter: gc.ShelterFull})
-		temp, err := AmbientTemperatureAt(world, 0, 0)
-		require.NoError(t, err)
-		return temp
-	}()
-
-	assert.Greater(t, indoor, outdoor, "冬でも屋内は屋外より暖かい退避先になる")
+// setStage は現ステージを切り替え、生成時と同じく基本気温を StageField へ写す。
+// 気温計算は StageField.BaseTemp を読むので、テストでも生成の手順を再現する
+func setStage(world w.World, key gc.StageKey) {
+	query.GetDungeon(world).CurrentStage = key
+	query.EnsureStageField(world, key).BaseTemp = dungeon.BaseTemperatureFor(key.Name)
 }
 
 func TestCalcBodyTempRate(t *testing.T) {
@@ -335,7 +147,7 @@ func TestTemperatureSystem_Update(t *testing.T) {
 	t.Run("寒い環境で体温が下がり続けると低体温に至る", func(t *testing.T) {
 		t.Parallel()
 		world := testutil.InitTestWorld(t)
-		query.GetDungeon(world).CurrentStage = gc.NewDungeonStage(coldDungeonName, 1) // 基本気温0度
+		setStage(world, gc.NewDungeonStage(coldDungeonName, 1)) // 基本気温0度
 
 		player, err := lifecycle.SpawnPlayer(world, consts.Coord[consts.Tile]{X: 0, Y: 0}, "ash")
 		require.NoError(t, err)
@@ -360,92 +172,12 @@ func TestTemperatureSystem_Update(t *testing.T) {
 	t.Run("存在しないダンジョン名の場合はエラーなし", func(t *testing.T) {
 		t.Parallel()
 		world := testutil.InitTestWorld(t)
-		query.GetDungeon(world).CurrentStage = gc.NewDungeonStage("存在しないダンジョン", 1)
+		setStage(world, gc.NewDungeonStage("存在しないダンジョン", 1))
 
 		sys := &TemperatureSystem{}
 		err := sys.Update(world)
 
 		assert.NoError(t, err)
-	})
-}
-
-func TestComfortableRange(t *testing.T) {
-	t.Parallel()
-
-	t.Run("断熱なしの場合の快適温度範囲", func(t *testing.T) {
-		t.Parallel()
-		lower, upper := ComfortableRange(Insulation{Cold: 0, Heat: 0})
-		assert.Equal(t, 11, lower)
-		assert.Equal(t, 30, upper)
-	})
-
-	t.Run("耐寒ありの場合は下限が下がる", func(t *testing.T) {
-		t.Parallel()
-		lower, upper := ComfortableRange(Insulation{Cold: 10, Heat: 0})
-		assert.Equal(t, 1, lower)
-		assert.Equal(t, 30, upper)
-	})
-
-	t.Run("耐熱ありの場合は上限が上がる", func(t *testing.T) {
-		t.Parallel()
-		lower, upper := ComfortableRange(Insulation{Cold: 0, Heat: 10})
-		assert.Equal(t, 11, lower)
-		assert.Equal(t, 40, upper)
-	})
-
-	t.Run("両方ありの場合", func(t *testing.T) {
-		t.Parallel()
-		lower, upper := ComfortableRange(Insulation{Cold: 15, Heat: 5})
-		assert.Equal(t, -4, lower)
-		assert.Equal(t, 35, upper)
-	})
-}
-
-func TestCalculateEquippedInsulation(t *testing.T) {
-	t.Parallel()
-
-	t.Run("装備なしの場合は全て0", func(t *testing.T) {
-		t.Parallel()
-		world := testutil.InitTestWorld(t)
-		player, err := lifecycle.SpawnPlayer(world, consts.Coord[consts.Tile]{X: 0, Y: 0}, "ash")
-		require.NoError(t, err)
-
-		insulation := CalculateEquippedInsulation(world, player)
-		assert.Equal(t, 0, insulation.Cold)
-		assert.Equal(t, 0, insulation.Heat)
-	})
-
-	t.Run("装備の断熱値が合算される", func(t *testing.T) {
-		t.Parallel()
-		world := testutil.InitTestWorld(t)
-		player, err := lifecycle.SpawnPlayer(world, consts.Coord[consts.Tile]{X: 0, Y: 0}, "ash")
-		require.NoError(t, err)
-
-		// 胴体装備（耐寒10, 耐熱5）
-		armor := world.ECS.NewEntity()
-		world.Components.Wearable.Add(armor, &gc.Wearable{
-			EquipmentCategory: gc.EquipmentTorso,
-			InsulationCold:    10,
-			InsulationHeat:    5,
-		})
-		world.Components.LocationEquipped.Add(armor, &gc.LocationEquipped{
-			Owner: player,
-		})
-
-		// 頭装備（耐寒3, 耐熱2）
-		helmet := world.ECS.NewEntity()
-		world.Components.Wearable.Add(helmet, &gc.Wearable{
-			EquipmentCategory: gc.EquipmentHead,
-			InsulationCold:    3,
-			InsulationHeat:    2,
-		})
-		world.Components.LocationEquipped.Add(helmet, &gc.LocationEquipped{
-			Owner: player,
-		})
-
-		insulation := CalculateEquippedInsulation(world, player)
-		assert.Equal(t, 13, insulation.Cold)
-		assert.Equal(t, 7, insulation.Heat)
 	})
 }
 
@@ -495,48 +227,12 @@ func TestLogTemperatureChange(t *testing.T) {
 	})
 }
 
-func TestHeatSourceWarmthAt_距離に応じて減衰し半径外は無視する(t *testing.T) {
-	t.Parallel()
-	world := testutil.InitTestWorld(t)
-
-	addHeatSource := func(x, y, radius consts.Tile, warmth float64) {
-		e := world.ECS.NewEntity()
-		world.Components.GridElement.Add(e, &gc.GridElement{Coord: consts.Coord[consts.Tile]{X: x, Y: y}})
-		// 暖房は HeatSource だけで決まる。Burning が無くても、電熱のように暖房になる
-		world.Components.HeatSource.Add(e, &gc.HeatSource{Radius: radius, Warmth: warmth})
-	}
-	addHeatSource(5, 5, 2, 0.6)
-	addHeatSource(20, 20, 1, 9.9) // 遠く、どの検証点からも圏外
-
-	// 源泉は満額、距離1は 2/3、距離2は 1/3 と線形に減衰する
-	assert.InDelta(t, 0.6, heatSourceWarmthAt(world, 5, 5), 1e-9)
-	assert.InDelta(t, 0.4, heatSourceWarmthAt(world, 6, 5), 1e-9)
-	assert.InDelta(t, 0.2, heatSourceWarmthAt(world, 7, 5), 1e-9)
-	assert.InDelta(t, 0.0, heatSourceWarmthAt(world, 8, 5), 1e-9)
-}
-
-func TestHeatSourceWarmthAt_複数の熱源を加算する(t *testing.T) {
-	t.Parallel()
-	world := testutil.InitTestWorld(t)
-
-	addHeatSource := func(x, y, radius consts.Tile, warmth float64) {
-		e := world.ECS.NewEntity()
-		world.Components.GridElement.Add(e, &gc.GridElement{Coord: consts.Coord[consts.Tile]{X: x, Y: y}})
-		// 暖房は HeatSource だけで決まる。Burning が無くても、電熱のように暖房になる
-		world.Components.HeatSource.Add(e, &gc.HeatSource{Radius: radius, Warmth: warmth})
-	}
-	addHeatSource(5, 5, 1, 0.5)
-	addHeatSource(6, 6, 2, 0.3)
-
-	assert.InDelta(t, 0.7, heatSourceWarmthAt(world, 5, 5), 1e-9)
-}
-
 func TestTemperatureSystem_Update_熱源のそばは体温の低下が緩む(t *testing.T) {
 	t.Parallel()
 
 	run := func(withBonfire bool) float64 {
 		world := testutil.InitTestWorld(t)
-		query.GetDungeon(world).CurrentStage = gc.NewDungeonStage(coldDungeonName, 1)
+		setStage(world, gc.NewDungeonStage(coldDungeonName, 1))
 		player, err := lifecycle.SpawnPlayer(world, consts.Coord[consts.Tile]{X: 5, Y: 5}, "ash")
 		require.NoError(t, err)
 		if withBonfire {
@@ -559,7 +255,7 @@ func TestTemperatureSystem_Update_熱源のそばは体温の低下が緩む(t *
 func TestBodyTempRate_寒い環境では負になる(t *testing.T) {
 	t.Parallel()
 	world := testutil.InitTestWorld(t)
-	query.GetDungeon(world).CurrentStage = gc.NewDungeonStage(coldDungeonName, 1)
+	setStage(world, gc.NewDungeonStage(coldDungeonName, 1))
 	player, err := lifecycle.SpawnPlayer(world, consts.Coord[consts.Tile]{X: 0, Y: 0}, "ash")
 	require.NoError(t, err)
 
@@ -569,7 +265,7 @@ func TestBodyTempRate_寒い環境では負になる(t *testing.T) {
 func TestBodyTempRate_冷えた体は熱源のそばで温まる(t *testing.T) {
 	t.Parallel()
 	world := testutil.InitTestWorld(t)
-	query.GetDungeon(world).CurrentStage = gc.NewDungeonStage(coldDungeonName, 1)
+	setStage(world, gc.NewDungeonStage(coldDungeonName, 1))
 	player, err := lifecycle.SpawnPlayer(world, consts.Coord[consts.Tile]{X: 5, Y: 5}, "ash")
 	require.NoError(t, err)
 	world.Components.HealthStatus.Get(player).BodyTempOffset = -3.0
@@ -586,7 +282,7 @@ func TestBodyTempRate_冷えた体は熱源のそばで温まる(t *testing.T) {
 func TestBodyTempRate_平熱以上では熱源が効かない(t *testing.T) {
 	t.Parallel()
 	world := testutil.InitTestWorld(t)
-	query.GetDungeon(world).CurrentStage = gc.NewDungeonStage("Debug town", 1)
+	setStage(world, gc.NewDungeonStage("Debug town", 1))
 	player, err := lifecycle.SpawnPlayer(world, consts.Coord[consts.Tile]{X: 5, Y: 5}, "ash")
 	require.NoError(t, err)
 
@@ -599,7 +295,7 @@ func TestBodyTempRate_平熱以上では熱源が効かない(t *testing.T) {
 func TestBodyTempRate_外因が無ければ平熱へ戻る(t *testing.T) {
 	t.Parallel()
 	world := testutil.InitTestWorld(t)
-	query.GetDungeon(world).CurrentStage = gc.NewDungeonStage("Debug town", 1)
+	setStage(world, gc.NewDungeonStage("Debug town", 1))
 	player, err := lifecycle.SpawnPlayer(world, consts.Coord[consts.Tile]{X: 5, Y: 5}, "ash")
 	require.NoError(t, err)
 	hs := world.Components.HealthStatus.Get(player)
